fix: strange test error
