+++ src/Form/constants.js
@@ -4,4 +4,14 @@ const FORM_CONTROL_SIZES = {
   LARGE: 'lg',
 };
 
-export { FORM_CONTROL_SIZES };
+const FORM_TEXT_TYPES = {
+  DEFAULT: 'default',
+  VALID: 'valid',
+  INVALID: 'invalid',
+  WARNING: 'warning',
+  CRITERIA_EMPTY: 'criteria-empty',
+  CRITERIA_VALID: 'criteria-valid',
+  CRITERIA_INVALID: 'criteria-invalid',
+};
+
+export { FORM_CONTROL_SIZES, FORM_TEXT_TYPES };
